Remove bad crackling when changing the volume through MIDI
From the GUI, the volume only changes at the start of a block.

From JACK MIDI, the volume can also change within a block. jack_mixer then
neatly makes sure the volume change is applied at exactly the right sample
in the buffer.

There was a bug in this code that introduced a bad crackle, which is fixed in
this commit. There's still a small crackle left, but this is another issue,
and also audible when changing the volume through the GUI.

Related to:
- https://gna.org/support/?2533
- https://gna.org/bugs/?15947