Don't re-create balance sliders when turning in/off custom widgets
The channel balance slider widget has no custom variant, so no
need to re-create it, when switching from/to custom widgets.

Previously it was re-created but not re-added to its parent,
whcih caused it to vanish. Make sure it is re-added to its parent
when it is re-created, just in case.

Fixes: #77

Signed-off-by: Christopher Arndt <chris@chrisarndt.de>