Sync output channel names with inputs on update
Whenever an output channel name is changed we need to make sure all
input channels have the correctly updated output channel name. This
sync is already done for channel colors, but as the name was only set
in the parent class, whereas the update already happens in
OutputChannelPropertiesDialog class. To make sure the name is properly
synced we duplicate setting the name and set it before all input
channels are synced with the output channel currently being altered.